feat(vm): add halo2 support
