+++ msg/msg.go
@@ -13,9 +13,3 @@ type MsgKey string
 func (m *Msg) Key() MsgKey {
 	return MsgKey(fmt.Sprintf("%s:%s", m.ProjectID, m.ProjectVersion))
 }
-
-type FetchStrategy string
-
-const (
-	FIFO FetchStrategy = "fifo"
-)
