+++ msg/types/types.go
@@ -1,0 +1,7 @@
+package types
+
+type FetchStrategy string
+
+const (
+	FIFO FetchStrategy = "fifo"
+)
