--- output/type.go
@@ -1,8 +1,0 @@
-package output
-
-type Type string
-
-const (
-	Stdout       Type = "stdout"
-	IoTeXTestNet Type = "iotex-testnet"
-)
