+++ output/types/types.go
@@ -1,0 +1,8 @@
+package types
+
+type Type string
+
+const (
+	Stdout       Type = "stdout"
+	IoTeXTestNet Type = "iotex-testnet"
+)
