--- vm/instance/instance.go
@@ -1,12 +1,0 @@
-package instance
-
-import (
-	"context"
-
-	"github.com/machinefi/w3bstream-mainnet/msg"
-)
-
-type Instance interface {
-	Execute(ctx context.Context, msg *msg.Msg) ([]byte, error)
-	Release()
-}
