--- vm/risc0/server.proto
@@ -1,26 +1,0 @@
-syntax = "proto3";
-package vm_runtime;
-option go_package = "pkg/runtime/risc0";
-
-service VmRuntime {
-    rpc Create(CreateRequest) returns (CreateResponse);
-    rpc ExecuteOperator(ExecuteRequest) returns (ExecuteResponse);
-}
-
-message CreateRequest {
-    string project = 1;
-    bytes content = 2;
-    string expParam = 3;
-}
-
-message CreateResponse {
-}
-
-message ExecuteRequest {
-    string project = 1;
-    string param = 2;
-}
-
-message ExecuteResponse {
-    bytes result = 1;
-}
