+++ vm/server/server.proto
@@ -1,0 +1,26 @@
+syntax = "proto3";
+package vm_runtime;
+option go_package = "vm/server";
+
+service VmRuntime {
+    rpc Create(CreateRequest) returns (CreateResponse);
+    rpc ExecuteOperator(ExecuteRequest) returns (ExecuteResponse);
+}
+
+message CreateRequest {
+    string project = 1;
+    bytes content = 2;
+    string expParam = 3;
+}
+
+message CreateResponse {
+}
+
+message ExecuteRequest {
+    string project = 1;
+    string param = 2;
+}
+
+message ExecuteResponse {
+    bytes result = 1;
+}
