--- vm/type.go
@@ -1,8 +1,0 @@
-package vm
-
-type Type string
-
-const (
-	Risc0 Type = "risc0"
-	Halo2 Type = "halo2"
-)
