+++ vm/types/types.go
@@ -1,0 +1,8 @@
+package types
+
+type Type string
+
+const (
+	Risc0 Type = "risc0"
+	Halo2 Type = "halo2"
+)
